requirements_dev.txt
celery==4.1.1
Django==2.0.5
djangorestframework-filters==0.10.2
djangorestframework==3.8.2
drf-extensions==0.3.1
raven==6.9.0
faker=0.8.15